Proposal: Enforce formatting
Added updated codeStyle.xml using the one from logback as the most
recent and added missing items from newer versions of eclipse.
Added java formatter with activation turned off for the time being.

This is a proposal that the java code get properly enforced to be
formatted.  By removing the commented activation this will occur on each
build.  After the first format, it will not format again except in cases
where a build is occurring and formatting was not followed.  This is
rather clean after the first cut.  This plugin covers only the java code
at this time.  XML will still need done the old fashioned way.