Add Log4J method to MDC to workaround use in wild
Adding a fix for #84 (http://bugzilla.slf4j.org/show_bug.cgi?id=84) that will allow downstream projects that incorrectly depend on Log4J methods that are not part of the public API to use this bridge. It is marked as deprecated to raise a compiler warning in these projects until they release fixes. For example, see http://jira.pentaho.com/browse/MONDRIAN-1171